requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
python-dotenv==1.1.1
supabase==2.16.0
httpx==0.28.1
requests==2.32.3

